+++ mealz-shared-analytics/src/commonMain/kotlin/mealz/ai/EventService.kt
@@ -160,7 +160,7 @@ object EventService : EventSender {
         PlausibleDestinations.PLANNER_ONBOARDING_UNDERSTOOD_CLICK.plausiblePath to propsOf(PlatformList("type"), PlatformList()),
         PlausibleDestinations.PLANNER_ONBOARDING_CLOSE_CLICK.plausiblePath to propsOf(PlatformList("type", "step"), PlatformList()),
 
-        // -------------------------------- PLANNER HELP ----------------------------------------
+        // -------------------------------- PLANNER HELP MODAL ----------------------------------------
         PlausibleDestinations.PLANNER_HELP_QUESTION_MARK_CLICK.plausiblePath to propsOf(PlatformList("from"), PlatformList()),
         PlausibleDestinations.PLANNER_HELP_DISPLAY.plausiblePath to propsOf(PlatformList("step"), PlatformList()),
         PlausibleDestinations.PLANNER_HELP_NEXT_CLICK.plausiblePath to propsOf(PlatformList("step"), PlatformList()),