+++ mealz-shared-analytics/src/commonMain/kotlin/mealz/ai/Init.kt
@@ -5,8 +5,8 @@ import kotlin.js.JsName
 
 @JsExport
 @JsName("initSharedAnalytics")
-fun initSharedAnalytics(domain: String, version: String, onEmit: onEmitFunction) {
-    SharedAnalytics.initSharedAnalytics(domain, version, onEmit)
+fun initSharedAnalytics(domain: String, version: String, onEmit: onEmitFunction, environment: String = "prod") {
+    SharedAnalytics.initSharedAnalytics(domain, version, onEmit, environment)
 }
 
 @JsExport