+++ mealz-shared-analytics/src/commonMain/kotlin/mealz/ai/SharedAnalytics.kt
@@ -5,6 +5,6 @@ import ai.mealz.analytics.utils.PlatformMap
 expect object SharedAnalytics {
     internal var abTestKey: String?
     internal var affiliate: String?
-    fun initSharedAnalytics(domain: String, version: String, onEmit: onEmitFunction)
+    fun initSharedAnalytics(domain: String, version: String, onEmit: onEmitFunction, environment: String = "prod")
     fun sendPlausibleRequest(plausiblePath: String, path: String, journey: String, plausibleProps: PlatformMap<String, String?>)
 }