+++ mealz-shared-analytics/src/jsMain/kotlin/mealz/ai/SharedAnalytics.js.kt
@@ -8,6 +8,7 @@ import org.w3c.fetch.RequestInit
 actual object SharedAnalytics : AbstractSharedAnalytics() {
 
     override fun sendRequest(event: PlausibleEvent) {
+        val analyticsUrl = getAnalyticsUrl()
         val headers = Headers()
         headers.set("Content-Type", "application/json")
         val body = JSON.stringify(event)
@@ -17,14 +18,14 @@ actual object SharedAnalytics : AbstractSharedAnalytics() {
             body
         )
         window.fetch(PLAUSIBLE_URL, init)
-        window.fetch(MEALZ_ANALYTICS_URL, init)
+        window.fetch(analyticsUrl, init)
     }
 
     actual fun sendPlausibleRequest(plausiblePath: String, path: String, journey: String, plausibleProps: PlatformMap<String, String?>) {
         this.buildAndSendPlausibleRequest(plausiblePath, path, journey, plausibleProps)
     }
 
-    actual fun initSharedAnalytics(domain: String, version: String, onEmit: onEmitFunction) {
-        this.init(domain, version, onEmit)
+    actual fun initSharedAnalytics(domain: String, version: String, onEmit: onEmitFunction, environment: String) {
+        this.init(domain, version, onEmit, environment)
     }
 }